+++ Modules/core/Controller/CorespaceadminController.php
@@ -114,6 +114,7 @@ class CorespaceadminController extends CoresecureController {
         $form->addTextArea("description", CoreTranslator::Description($lang), false, $space["description"]);
         $form->addText("contact", CoreTranslator::Contact($lang), true, $space["contact"]);
         $form->addText("support", CoreTranslator::Support($lang), false, $space["support"]);
+        $form->addText("termsofuse", CoreTranslator::Policy($lang), false, $space['termsofuse'] ?? '');
 
         $modelUser = new CoreUser();
         $users = $modelUser->getActiveUsers("name");
@@ -223,6 +224,9 @@ class CorespaceadminController extends CoresecureController {
             $modelSpace->setDescription($id, $this->request->getParameter("description"));
             $modelSpace->setAdmins($id, $this->request->getParameter("admins"));
             $modelSpace->setDeactivate($id, $this->request->getParameter('user_desactivate'));
+
+            $termsofuse = $this->request->getParameterNoException("termsofuse");
+            $modelSpace->setTermsOfUse($id, $termsofuse);
             $modelSpace->setOnDeactivate($id, $this->request->getParameter('on_user_deactivate'));
             
             // upload image