+++ Modules/core/Model/CoreTranslator.php
@@ -1939,7 +1939,14 @@ class CoreTranslator {
         return "Join requested..."; 
     }
 
-    public static function JoinRequestEmail($login, $spaceName, $userEmail, $userFullName, $lang, $organization = null, $team = null) {
+    public static function JoinWhy($lang){
+        if ($lang == "fr") {
+            return "Qui êtes vous, de quoi avez vous besoin?";
+        }
+        return "Who are you, and what do you need?"; 
+    }
+
+    public static function JoinRequestEmail($login, $spaceName, $userEmail, $userFullName, $lang, $organization = null, $team = null, $comment='') {
         if ($lang == "fr") {
             $message = "Bonjour, <br><br>" . $userFullName . " demande à rejoindre votre espace " . $spaceName. " sur Platform-Manager";
             $message .= ("<br>Login : " . $login);
@@ -1950,6 +1957,7 @@ class CoreTranslator {
             if ($team) {
                 $message .= ("<br>Equipe : " . $team);
             }
+            $message .= ('<br><br>Commentaire: ' . $comment);
             return $message;
         }
         $message = "Hi, <br><br>" . $userFullName . " requests to join your space " . $spaceName. " on Platform-Manager";
@@ -1961,6 +1969,7 @@ class CoreTranslator {
         if ($team) {
             $message .= ("<br>Team: " . $team);
         }
+        $message .= ('<br><br>Comment: '.$comment);
         return $message;
     }
 
@@ -2006,6 +2015,13 @@ class CoreTranslator {
             return "Email de support";
         }
         return "Support email";
+    }
+
+    public static function Policy($lang = "") {
+        if ($lang == "fr") {
+            return "Charte d'utilisation";
+        }
+        return "Usage policy";
     }
 
     public static function MailSubjectPrefix($spaceName = "") {